Checked in final changes for taxonomy_term_reference autocomplete field
Now the autocomplete field can populate existing value or values and can add or remove values. Upon submission, the values are stored in form_state and subsequently updated to the field of the node. A preliminary testing for an existing node is working. Haven't test creating a new node. There must be bugs.